fix redeem from airdrop
